feat(internal.services): make da shit bete

+++ internal/services/table.go
@@ -1,13 +1,19 @@
 package services
 
 import (
+	"errors"
+
 	"github.com/mksmstpck/restoracio/internal/models"
+	"github.com/mksmstpck/restoracio/utils"
 	"github.com/patrickmn/go-cache"
 	"github.com/pborman/uuid"
 	log "github.com/sirupsen/logrus"
 )
 
 func (s *Services) TableCreateService(table models.Table, admin models.Admin) (models.Table, error) {
+	if admin.Restaurant == nil {
+		return models.Table{}, errors.New("create restaurant first")
+	}
 	table.RestaurantID = admin.Restaurant.ID
 	res, err := s.db.Table.CreateOne(s.ctx, table)
 	if err != nil {
@@ -37,7 +43,22 @@ func (s *Services) TableGetAllInRestaurantService(id uuid.UUID) ([]models.Table,
 	return tables, nil
 }
 
-func (s *Services) TableUpdateService(table models.Table) error {
+func (s *Services) TableUpdateService(table models.Table, admin models.Admin) error {
+	if admin.Restaurant == nil {
+		log.Info("create restaurant first")
+		return errors.New(utils.ErrRestaurantNotFound)
+	}
+	if admin.Restaurant.Tables == nil {
+		log.Info("create tables first")
+		return errors.New(utils.ErrTableNotFound)
+	}
+	table.RestaurantID = admin.Restaurant.ID
+
+	if !tableExists(admin.Restaurant.Tables, table.ID) {
+		log.Info("table not found")
+		return errors.New("table not found")
+	}
+
 	err := s.db.Table.UpdateOne(s.ctx, table)
 	if err != nil {
 		log.Info("TableUpdate: ", err)
@@ -47,7 +68,19 @@ func (s *Services) TableUpdateService(table models.Table) error {
 	return nil
 }
 
-func (s *Services) TableDeleteService(id uuid.UUID) error {
+func (s *Services) TableDeleteService(id uuid.UUID, admin models.Admin) error {
+	if admin.Restaurant == nil {
+		log.Info("create restaurant first")
+		return errors.New(utils.ErrRestaurantNotFound)
+	}
+	if admin.Restaurant.Tables == nil {
+		log.Info("create tables first")
+		return errors.New(utils.ErrTableNotFound)
+	}
+	if !tableExists(admin.Restaurant.Tables, id.String()) {
+		log.Info("table not found")
+		return errors.New("table not found")
+	}
 	err := s.db.Table.DeleteOne(s.ctx, id)
 	if err != nil {
 		log.Info("TableDelete: ", err)
@@ -55,4 +88,13 @@ func (s *Services) TableDeleteService(id uuid.UUID) error {
 	}
 	log.Info("table deleted")
 	return nil
+}
+
+func tableExists(tables []*models.Table, id string) bool {
+	for _, table := range tables {
+		if table.ID == id {
+			return true
+		}
+	}
+	return false
 }
